requirements.txt
pymupdf
pillow

transformers
torch

google-auth
google-api-python-client
